requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
beautifulsoup4==4.6.0
boto3==1.4.1
botocore==1.4.93          # via boto3, s3transfer
cachetools==2.0.1         # via zenpy
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
contextlib2==0.5.5        # via raven
directory-api-client==9.16.0
directory-client-core==0.2.1  # via directory-api-client, directory-cms-client
directory-cms-client==2.0.0
directory_components==2.2.0
directory_constants==6.7.0
directory_validators==4.5.0
django-formtools==2.0
django-storages==1.5.1
django==1.11.10
django_recaptcha==1.2.0
djangorestframework==3.7.7  # via sigauth
docutils==0.14            # via botocore
export-elements==0.11.1
gunicorn==19.5.0
idna==2.6                 # via requests
jmespath==0.9.3           # via boto3, botocore
jsonschema==2.6.0
mohawk==0.3.4             # via sigauth
monotonic==1.4            # via directory-client-core
olefile==0.44             # via directory-validators, pillow
phonenumberslite==7.7.4   # via directory-validators
pillow==4.2.1             # via directory-validators
python-dateutil==2.6.1    # via botocore, zenpy
pytz==2017.2              # via directory-constants, directory-validators, django
raven==5.19.0
redis==2.10.5
requests==2.18.4
s3transfer==0.1.13        # via boto3
sigauth==2.1.2            # via directory-client-core
six==1.11.0               # via mohawk, python-dateutil
sorl-thumbnail==12.3
urllib3==1.22             # via requests
waitress==0.8.9
whitenoise==3.1
zenpy==1.1.10
